Move subjects to @store/subjects It's much quicker to type.
