+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/InteractionLayer/components/TranscribedLines/TranscribedLines.spec.js
@@ -1,10 +1,10 @@
 import { shallow } from 'enzyme'
 import sinon from 'sinon'
 import React from 'react'
-import TranscriptionReductions from '@store/SubjectStore/SubjectType/subjects/Subject/TranscriptionReductions'
+import TranscriptionReductions from '@store/subjects/Subject/TranscriptionReductions'
 import * as tasks from '@plugins/tasks'
 import { TranscribedLines } from './TranscribedLines'
-import { reducedSubject } from '@store/SubjectStore/SubjectType/subjects/Subject/TranscriptionReductions/mocks'
+import { reducedSubject } from '@store/subjects/Subject/TranscriptionReductions/mocks'
 import { TranscriptionLine } from '@plugins/drawingTools/components'
 import ConsensusPopup from './components/ConsensusPopup'
 import { expect } from 'chai'
