+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/InteractionLayer/components/TranscribedLines/TranscribedLinesConnector.spec.js
@@ -1,7 +1,7 @@
 import { shallow } from 'enzyme'
 import React from 'react'
-import TranscriptionReductions from '@store/SubjectStore/SubjectType/subjects/Subject/TranscriptionReductions'
-import { reducedSubject } from '@store/SubjectStore/SubjectType/subjects/Subject/TranscriptionReductions/mocks'
+import TranscriptionReductions from '@store/subjects/Subject/TranscriptionReductions'
+import { reducedSubject } from '@store/subjects/Subject/TranscriptionReductions/mocks'
 import TranscribedLinesConnector from './TranscribedLinesConnector'
 import TranscribedLines from './TranscribedLines'
