+++ packages/lib-classifier/src/store/subjects/ImageAndTextSubject/ImageAndTextSubject.spec.js
