+++ packages/lib-classifier/src/store/subjects/SingleImageSubject/SingleImageSubject.spec.js
