+++ packages/lib-classifier/src/store/subjects/SingleImageSubject/index.js
