+++ packages/lib-classifier/src/store/subjects/SingleTextSubject/SingleTextSubject.js
