+++ packages/lib-classifier/src/store/subjects/SingleTextSubject/index.js
