+++ packages/lib-classifier/src/store/subjects/SingleVideoSubject/SingleVideoSubject.spec.js
