+++ packages/lib-classifier/src/store/subjects/Subject/TranscriptionReductions/mocks/reducedEmptySubject.js
