+++ packages/lib-classifier/src/store/subjects/TextSubject/TextSubject.js
