+++ packages/lib-classifier/src/store/subjects/TextSubject/index.js
